refactor #212 活動記録関連テストの整理
