+++ backend/test/routers/test_activity_actual.py
@@ -1,0 +1,291 @@
+from test.helpers.activity import (
+    test_date,
+    setup_target_time,
+    setup_actual_time,
+    setup_finish_activity,
+    setup_monthly_income
+)
+from app.error_codes import NotFoundCode, BadRequestCode, ConflictCode
+
+
+def test_register_actual(client, get_resource_owner_headers):
+    """ 1つの活動時間を登録した場合 """
+    setup_monthly_income(client, get_resource_owner_headers)
+    setup_target_time(client, get_resource_owner_headers)
+    # 活動時間を登録
+    data = {
+        "activities": [
+            {"date": test_date, "actual_time": 5.0},
+        ]
+    }
+    response = client.put("/activities/actual",
+                          json=data,
+                          headers=get_resource_owner_headers)
+    assert response.status_code == 200
+    assert response.json() == {
+        "results": [
+            {
+                "date": test_date,
+                "result": "success",
+                "actual_time": 5.0,
+                "reason": None
+            }
+        ]
+    }
+
+
+def test_register_multi_actual(client, get_resource_owner_headers):
+    """ 複数の活動時間を登録した場合 """
+    setup_monthly_income(client, get_resource_owner_headers)
+    # 目標時間を複数登録
+    data = {
+        "activities": [
+            {"date": test_date, "target_time": 5.0},
+            {"date": "2024-5-6", "target_time": 6.0},
+            {"date": "2024-5-7", "target_time": 7.0}
+        ]
+    }
+    client.post("/activities/target",
+                json=data,
+                headers=get_resource_owner_headers)
+    # 活動時間を登録
+    data = {
+        "activities": [
+            {"date": test_date, "actual_time": 5.0},
+            {"date": "2024-5-6", "actual_time": 6.0},
+            {"date": "2024-5-7", "actual_time": 7.0}
+        ]
+    }
+    response = client.put("/activities/actual",
+                          json=data,
+                          headers=get_resource_owner_headers)
+    assert response.status_code == 200
+    assert response.json() == {
+        "results": [
+            {
+                "date": test_date,
+                "result": "success",
+                "actual_time": 5.0,
+                "reason": None
+            },
+            {
+                "date": "2024-5-6",
+                "result": "success",
+                "actual_time": 6.0,
+                "reason": None
+            },
+            {
+                "date": "2024-5-7",
+                "result": "success",
+                "actual_time": 7.0,
+                "reason": None
+            }
+        ]
+    }
+
+
+def test_register_actual_with_partial_error(client, get_resource_owner_headers):
+    """ 目標時間が登録されていないものが含まれる場合 """
+    setup_monthly_income(client, get_resource_owner_headers)
+    setup_target_time(client, get_resource_owner_headers)
+    data = {
+        "activities": [
+            {"date": test_date, "actual_time": 5.0},
+            {"date": "2024-5-11", "actual_time": 5.0}
+        ]
+    }
+    response = client.put("/activities/actual",
+                          json=data,
+                          headers=get_resource_owner_headers)
+    assert response.status_code == 200
+    assert response.json() == {
+        "results": [
+            {
+                "date": test_date,
+                "result": "success",
+                "actual_time": 5.0,
+                "reason": None
+            },
+            {
+                "date": "2024-5-11",
+                "result": "error",
+                "actual_time": None,
+                "reason": NotFoundCode.ACTIVITY_NOT_FOUND
+            }
+        ]
+    }
+
+
+def test_register_actual_with_all_errors(client, get_resource_owner_headers):
+    """ 全てのリクエストがエラーになる場合 """
+    data = {
+        "activities": [
+            {"date": test_date, "actual_time": 5.0},
+            {"date": "2024-5-11", "actual_time": 5.0}
+        ]
+    }
+    response = client.put("/activities/actual",
+                          json=data,
+                          headers=get_resource_owner_headers)
+    assert response.status_code == 400
+    assert response.json() == {
+        "code": BadRequestCode.BULK_ACTIVITY_OPERATION_FAILED,
+        "results": [
+            {
+                "date": test_date,
+                "result": "error",
+                "actual_time": None,
+                "reason": NotFoundCode.SALARY_NOT_FOUND
+            },
+            {
+                "date": "2024-5-11",
+                "result": "error",
+                "actual_time": None,
+                "reason": NotFoundCode.SALARY_NOT_FOUND
+            }
+        ]
+    }
+
+
+def test_register_actual_with_invalid_hour(client, get_resource_owner_headers):
+    """ 時間を1x.0or5、もしくはx.0or5の形で入力していない場合 """
+    setup_target_time(client, get_resource_owner_headers)
+    data = {
+        "activities": [
+            {"date": "2024-5-10", "actual_time": 5.2}
+        ]
+    }
+    response = client.put("/activities/actual",
+                          json=data,
+                          headers=get_resource_owner_headers)
+    assert response.status_code == 422
+    assert response.json() == {
+        "code": "VALIDATION_ERROR",
+        "errors": [
+            {
+                "code": "INVALID_VALUE",
+                "field": "actual_time"
+            }
+        ]
+    }
+
+
+def test_register_actual_after_finish(client, get_resource_owner_headers):
+    """ 活動を終了した日の活動時間を更新しようとした場合 """
+    setup_monthly_income(client, get_resource_owner_headers)
+    setup_target_time(client, get_resource_owner_headers)
+    setup_actual_time(client, get_resource_owner_headers)
+    setup_finish_activity(client, get_resource_owner_headers)
+    data = {
+        "activities": [
+            {"date": test_date, "actual_time": 5.0}
+        ]
+    }
+    response = client.put("/activities/actual",
+                          json=data,
+                          headers=get_resource_owner_headers)
+    assert response.status_code == 400
+    assert response.json() == {
+        "code": BadRequestCode.BULK_ACTIVITY_OPERATION_FAILED,
+        "results": [
+            {
+                "date": test_date,
+                "result": "error",
+                "actual_time": None,
+                "reason": ConflictCode.ACTIVITY_ALREADY_FINISHED
+            }
+        ]
+    }
+
+
+def test_register_multi_actual_with_invalid_hour(client, get_resource_owner_headers):
+    """ 複数の活動時間を登録する際に不正なデータが含まれている場合 """
+    setup_monthly_income(client, get_resource_owner_headers)
+    # 活動時間を登録
+    data = {
+        "activities": [
+            {"date": test_date, "actual_time": 15.0}  # 上限を超える活動時間
+        ]
+    }
+    response = client.put("/activities/actual",
+                          json=data,
+                          headers=get_resource_owner_headers)
+    assert response.status_code == 422
+    assert response.json() == {
+        "code": "VALIDATION_ERROR",
+        "errors": [
+            {
+                "code": "INVALID_VALUE",
+                "field": "actual_time"
+            }
+        ]
+    }
+
+
+def test_register_multi_actual_with_invalid_year(client, get_resource_owner_headers):
+    setup_monthly_income(client, get_resource_owner_headers)
+    data = {
+        "activities": [
+            {"date": "20240-5-5", "actual_time": 5.0}
+        ]
+    }
+
+    response = client.put("/activities/actual",
+                          json=data,
+                          headers=get_resource_owner_headers)
+    assert response.status_code == 422
+    assert response.json() == {
+        "code": "VALIDATION_ERROR",
+        "errors": [
+            {
+                "code": "INVALID_YEAR",
+                "field": "year"
+            }
+        ]
+    }
+
+
+def test_register_multi_actual_with_invalid_month(client, get_resource_owner_headers):
+    setup_monthly_income(client, get_resource_owner_headers)
+    data = {
+        "activities": [
+            {"date": "2025-50-5", "actual_time": 5.0}
+        ]
+    }
+
+    response = client.put("/activities/actual",
+                          json=data,
+                          headers=get_resource_owner_headers)
+    assert response.status_code == 422
+    assert response.json() == {
+        "code": "VALIDATION_ERROR",
+        "errors": [
+            {
+                "code": "INVALID_MONTH",
+                "field": "month"
+            }
+        ]
+    }
+
+
+def test_register_multi_actual_with_invalid_date(client, get_resource_owner_headers):
+    setup_monthly_income(client, get_resource_owner_headers)
+    data = {
+        "activities": [
+            {"date": "2024-5-50", "actual_time": 5.0}
+        ]
+    }
+
+    response = client.put("/activities/actual",
+                          json=data,
+                          headers=get_resource_owner_headers)
+    assert response.status_code == 422
+    assert response.json() == {
+        "code": "VALIDATION_ERROR",
+        "errors": [
+            {
+                "code": "INVALID_DATE",
+                "field": "date"
+            }
+        ]
+    }
